release/英语学习助手_Windows/backend/requirements.txt
# FastAPI 核心
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
pydantic-settings>=2.1.0

# 数据库
sqlalchemy>=2.0.25
aiosqlite>=0.19.0

# 认证
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# AI集成
openai>=1.10.0
anthropic>=0.18.0

# 工具库
python-dotenv>=1.0.0
httpx>=0.26.0

# WebSocket和缓存
websockets>=12.0
redis>=5.0.1

# 开发工具 (可选)
# pytest>=7.4.4
# pytest-asyncio>=0.23.3
